Move the interfaces to manage the state to leshan-core package
Move the interfaces `Destroyable`, `Startable` and `Stoppable` to
`leshan-core` package from `leshan-server-core`.
This commit aims to use these interfaces with `LwM2mObjectEnabler`
to control the state of that.

Ref: https://github.com/eclipse/leshan/issues/933#issuecomment-731036021
