+++ leshan-core/src/main/java/org/eclipse/leshan/core/Startable.java
@@ -13,11 +13,10 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.server;
+package org.eclipse.leshan.core;
 
-public interface Destroyable {
+public interface Startable {
 
-    /** Destroy the instances and frees all system resources. */
-    void destroy();
+    void start();
 
 }
