+++ leshan-core/src/main/java/org/eclipse/leshan/core/Stoppable.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.server;
+package org.eclipse.leshan.core;
 
 public interface Stoppable {
